Add Mobile UI Tab admin page and update routing
Introduces a new MobileUITab admin page for managing mobile app tabs, updates the sidebar and admin routes to include it, and refactors the Settings page to remove tab management logic. Also uncomments and enables the Events page logic.
