+++ frontend/web/churchlink/src/features/admin/pages/Events.tsx
@@ -1,42 +1,42 @@
 
-// // import { useAuth } from "@/lib/auth-context";\
-// import { useEffect, useState } from "react";
-// import { fetchEvents } from "@/helpers/EventsHelper";
-// import { fetchPermissions } from "@/helpers/PermissionsHelper";
-// import { ChurchEvent } from "@/shared/types/ChurchEvent";
-// import { AccountPermissions } from "@/shared/types/AccountPermissions";
-// import EventsTable from "@/features/admin/components/Events/EventsTable";
+// import { useAuth } from "@/lib/auth-context";\
+import { useEffect, useState } from "react";
+import { fetchEvents } from "@/helpers/EventsHelper";
+import { fetchPermissions } from "@/helpers/PermissionsHelper";
+import { ChurchEvent } from "@/shared/types/ChurchEvent";
+import { AccountPermissions } from "@/shared/types/AccountPermissions";
+import EventsTable from "@/features/admin/components/Events/EventsTable";
 
 
 
 
-// const Events = () => {
-//     const [events, setEvents] = useState<ChurchEvent[]>([]);
-//     const [perms, setPerms] = useState<AccountPermissions[]>([]);
-//     const [loading, setLoading] = useState(true);
+const Events = () => {
+    const [events, setEvents] = useState<ChurchEvent[]>([]);
+    const [perms, setPerms] = useState<AccountPermissions[]>([]);
+    const [loading, setLoading] = useState(true);
 
-//     const loadData = async () => {
-//         const eventsFromAPI = await fetchEvents();
-//         setEvents(eventsFromAPI);
-//         const permsFromAPI = await fetchPermissions();
-//         setPerms(permsFromAPI);
-//         setLoading(false);
-//     }
+    const loadData = async () => {
+        const eventsFromAPI = await fetchEvents();
+        setEvents(eventsFromAPI);
+        const permsFromAPI = await fetchPermissions();
+        setPerms(permsFromAPI);
+        setLoading(false);
+    }
 
-//     useEffect(() => {
-//         loadData();
-//     }, []);
+    useEffect(() => {
+        loadData();
+    }, []);
 
 
 
-//     if (loading) return <p>Loading Events...</p>;
+    if (loading) return <p>Loading Events...</p>;
 
-//     return (
-//         <div className="p-6">
-//             <h1 className="text-xl font-bold mb-4">Events Management</h1>
-//             <EventsTable data={events} permData={perms} onSave={loadData}></EventsTable>
-//         </div>
-//     );
-// };
+    return (
+        <div className="p-6">
+            <h1 className="text-xl font-bold mb-4">Events Management</h1>
+            <EventsTable data={events} permData={perms} onSave={loadData}></EventsTable>
+        </div>
+    );
+};
 
-// export default Events;
+export default Events;
